javascript / responsive
Javascript toegevoegd.
Website grotendeels responsive gemaakt

+++ helpers/edit_menu_db.php
@@ -6,7 +6,7 @@ $stmt->execute();
 $result = $stmt->fetchAll();
 
 ?>
-<table>
+<table class="edit-table">
     <th>
         <h2>Menu</h2>
     </th>
@@ -21,7 +21,7 @@ foreach($result as $value)
             <td>
             <?php  echo $value['Price'];?>
             </td>
-            <td class="test">
+            <td class="beschr">
                 <br>
             <?php  echo $value['Beschrijving'];?>
             </td>
@@ -29,12 +29,12 @@ foreach($result as $value)
                 <a href="CRUD/edit.php?id=<?php echo $value["Id"]; ?>">
                  <button class="btn-update"> <p><i class="fas fa-pen"></i> Update</p></button>   
                 </a>
-                <a href="CRUD/delete.php?id=<?php echo $value["Id"]; ?>">
-                 <button class="btn-delete"> <p><i class="fas fa-trash-alt"></i> Delete</p></button>   
+                <a href="CRUD/delete.php?id=<?php echo $value["Id"]; ?>" id="confirmClickActionElementId">
+                 <button class="btn-delete" onclick="return confirm('Weet u zeker dat u dit product wilt verwijderen?');" > <p><i class="fas fa-trash-alt"></i> Delete</p></button>   
                 </a>
             </td>
         </tr>
-    <?php
+        <?php
 }
 ?>
 </table>
